requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.37.0
peft>=0.9.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99

# Data processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.1

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
huggingface_hub>=0.19.0
rich>=13.5.0
wandb>=0.15.0  # Optional for experiment tracking

# For P-Tuning
einops>=0.6.0

# For QLoRA
optimum>=1.12.0

# For logging and visualization
tensorboard>=2.14.0
matplotlib>=3.7.0

# For common utilities
pathlib>=1.0.1
typing-extensions>=4.7.0
